+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.1.1 (working on it)
+
+major changes
+
+minor changes
+* static.py -> static_preview.py
+* 将 lega.draw 里面的 font_size 和 line_distance 的默认值
+  从绝对大小修改为与窗口宽度相关的相对大小
+* 将 FPS 的默认值从 30 改为 20
+* 重构 pygame_boy_advance.py 的代码
+* 一个空的 Magic Tower 界面
+
 ## v0.1.0
 
 * 一个结构完整且基本没有冗余的项目框架